Split community API views to smaller files
Split the experimental community API view code from a single large code
file to multiple smaller ones to make it easier to add more endpoints.
