+++ django/thunderstore/community/api/experimental/views/_utils.py
@@ -1,0 +1,44 @@
+from collections import OrderedDict
+
+from rest_framework.generics import ListAPIView
+from rest_framework.pagination import CursorPagination
+from rest_framework.response import Response
+
+
+class CustomCursorPagination(CursorPagination):
+    ordering = "-datetime_created"
+    results_name = "results"
+    page_size = 100
+
+    def get_paginated_response(self, data) -> Response:
+        return Response(
+            OrderedDict(
+                [
+                    (
+                        "pagination",
+                        OrderedDict(
+                            [
+                                ("next_link", self.get_next_link()),
+                                ("previous_link", self.get_previous_link()),
+                            ],
+                        ),
+                    ),
+                    (self.results_name, data),
+                ],
+            ),
+        )
+
+
+class CustomListAPIView(ListAPIView):
+    pagination_class = CustomCursorPagination
+    paginator: CustomCursorPagination
+
+    def list(self, request, *args, **kwargs):
+        queryset = self.filter_queryset(self.get_queryset())
+
+        page = self.paginate_queryset(queryset)
+        if page is None:
+            raise ValueError("Pagination not set")
+
+        serializer = self.get_serializer(page, many=True)
+        return self.get_paginated_response(serializer.data)
